FIX: Ajustes finais e correção de bugs.

+++ src/main/java/br/com/fabreum/AppProdutos/controller/ProdutoController.java
@@ -1,69 +1,104 @@
 package br.com.fabreum.AppProdutos.controller;
 
 import br.com.fabreum.AppProdutos.controller.dto.ProdutoDto;
+import br.com.fabreum.AppProdutos.controller.dto.ProdutoResponse;
 import br.com.fabreum.AppProdutos.model.Produtos;
 import br.com.fabreum.AppProdutos.repository.ProductRepository;
 import br.com.fabreum.AppProdutos.service.ProdutosService;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("api/v1/produtos/")
+@RequestMapping("api/produtos/")
+@Tag(name = "Produtos", description = "Endpoints for managing products")
 public class ProdutoController {
 
     private final ProductRepository produtosRepository;
     private final ProdutosService produtosService;
 
-    /**
-     * Cria um novo produto.
-     * Apenas usuários com o papel de ADMIN ou SELLER podem acessar este endpoint.
-     */
-    @SecurityRequirement(name = "bearerAuth")
+    @Operation(summary = "Cria um novo produto", description = "Adiciona produtos para venda.")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Produto criado com sucesso"),
+            @ApiResponse(responseCode = "403", description = "Acesso negado")
+    })
     @PostMapping("produto")
-    @PreAuthorize("hasRole('ADMIN') or hasRole('SELLER')")
-    public ResponseEntity<Produtos> criaProduto(@RequestBody ProdutoDto produtoDto) {
+    public ResponseEntity<ProdutoResponse> criaProduto(@RequestBody ProdutoDto produtoDto) {
         Produtos produto = new Produtos();
         produto.setName(produtoDto.getName());
         produto.setDescription(produtoDto.getDescription());
         produto.setPrice(produtoDto.getPrice());
-        produto.setSku("");
+        produto.setSku(produtoDto.getSku());
+        produto.setStockQuantity(produtoDto.getStockQuantity());
         Produtos saved = produtosRepository.save(produto);
-        return ResponseEntity.ok(saved);
+        return ResponseEntity.ok(new ProdutoResponse(
+                saved.getId(),
+                saved.getName(),
+                saved.getDescription(),
+                saved.getSku(),
+                saved.getPrice(),
+                saved.getCreatedAt(),
+                saved.getStockQuantity()
+        ));
+
     }
 
-    /**
-     * Lista todos os produtos.
-     * Acessível por qualquer usuário autenticado (ADMIN, SELLER, CUSTOMER).
-     */
+    @Operation(summary = "Lista todos os produtos", description = "Lista todos os produtos que existe no banco.")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Lista de produtos retornada com sucesso")
+    })
     @GetMapping
-    public ResponseEntity<List<Produtos>> listaProdutos() {
+    public ResponseEntity<List<ProdutoResponse>> listaProdutos() {
         List<Produtos> produtos = produtosRepository.findAll();
-        return ResponseEntity.ok(produtos);
+
+        List<ProdutoResponse> responseList = produtos.stream()
+                .map(prod -> new ProdutoResponse(
+                        prod.getId(),
+                        prod.getName(),
+                        prod.getDescription(),
+                        prod.getSku(),
+                        prod.getPrice(),
+                        prod.getCreatedAt(),
+                        prod.getStockQuantity()
+                ))
+                .toList();
+
+        return ResponseEntity.ok(responseList);
     }
 
-    /**
-     * Busca um produto pelo ID.
-     * Acessível por qualquer usuário autenticado.
-     */
+
+
+    @Operation(summary = "Busca um produto pelo ID", description = "Busca por um ID específico")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Produto encontrado com sucesso"),
+            @ApiResponse(responseCode = "404", description = "Produto não encontrado")
+    })
     @GetMapping("{id}")
-    public ResponseEntity<Produtos> listaProdutoPorId(@PathVariable Long id) {
+    public ResponseEntity<ProdutoResponse> listaProdutoPorId(@PathVariable Long id) {
         Produtos produto = produtosRepository.findById(id).orElseThrow();
-        return ResponseEntity.ok(produto);
+        return ResponseEntity.ok(new ProdutoResponse(
+                produto.getId(),
+                produto.getName(),
+                produto.getDescription(),
+                produto.getSku(),
+                produto.getPrice(),
+                produto.getCreatedAt(),
+                produto.getStockQuantity()
+        ));
     }
 
-    /**
-     * Exemplo de retorno de um Record.
-     * Acessível por qualquer usuário autenticado.
-     * @param id
-     * @return
-     */
+    @Operation(summary = "Exemplo de retorno de um Record", description = "Acessível por qualquer usuário autenticado.")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Produto DTO retornado com sucesso"),
+            @ApiResponse(responseCode = "404", description = "Produto não encontrado")
+    })
     @GetMapping("/dto/{id}")
     public ResponseEntity<ProdutoDto> listaProdutoDtoPorId(@PathVariable Long id) {
         Produtos produto = produtosRepository.findById(id).orElseThrow();
@@ -74,25 +109,44 @@ public class ProdutoController {
         return ResponseEntity.ok(produtoDto);
     }
 
-    /**
-     * Atualiza um produto existente.
-     * Apenas usuários com o papel de ADMIN ou SELLER podem acessar este endpoint.
-     * Uma lógica mais complexa poderia ser adicionada para que um SELLER só possa editar os próprios produtos.
-     */
+    @Operation(summary = "Atualiza um produto existente", description = "Altera/Edita produtos.")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Produto atualizado com sucesso"),
+            @ApiResponse(responseCode = "403", description = "Acesso negado"),
+            @ApiResponse(responseCode = "404", description = "Produto não encontrado")
+    })
     @PutMapping("atualiza/{id}")
-    @PreAuthorize("hasRole('ADMIN') or hasRole('SELLER')")
-    public ResponseEntity<Optional<Produtos>> atualizaProduto(@PathVariable Long id, @RequestBody ProdutoDto produto) {
-        final var produtoExistente = produtosService.atualizaProduto(id, produto);
-        return ResponseEntity.ok(produtoExistente);
+    public ResponseEntity<ProdutoResponse> atualizaProduto(@PathVariable Long id,
+                                                           @RequestBody ProdutoDto dto) {
+
+        Produtos atualizado = produtosService.atualizaProduto(id, dto)
+                .orElseThrow(() -> new RuntimeException("Produto não encontrado"));
+
+        ProdutoResponse response = new ProdutoResponse(
+                atualizado.getId(),
+                atualizado.getName(),
+                atualizado.getDescription(),
+                atualizado.getSku(),
+                atualizado.getPrice(),
+                atualizado.getCreatedAt(),
+                atualizado.getStockQuantity()
+        );
+
+        return ResponseEntity.ok(response);
     }
 
-    /**
-     * Deleta um produto.
-     * Apenas usuários com o papel de ADMIN podem acessar este endpoint.
-     */
+
+    @Operation(summary = "Deleta um produto", description = "Excluir um produto do banco pelo ID.")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "Produto deletado com sucesso"),
+            @ApiResponse(responseCode = "403", description = "Acesso negado")
+    })
     @DeleteMapping("{id}")
-    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Void> deletaProduto(@PathVariable Long id) {
+        if (!produtosRepository.existsById(id)) {
+            throw new RuntimeException("Produto não encontrado");
+        }
+
         produtosRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
